requirements.txt
# GitHub API integration
PyGithub>=1.59.0
requests>=2.31.0

# Google Gemini API
google-genai>=0.3.0

# Content processing
markdown>=3.5.0
nbformat>=5.9.0
jupyter-client>=8.0.0

# Configuration and utilities
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# Logging and utilities
colorama>=0.4.6
rich>=13.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0